feat(course): 수강신청 동시성제어 추가

+++ src/main/java/org/hyejoon/cuvcourse/domain/lecture/lectureView/dto/LectureViewResponse.java
@@ -1,4 +1,4 @@
 package org.hyejoon.cuvcourse.domain.lecture.lectureView.dto;
 
-public record LectureViewResponse (Long id, String lectureTitle, String professorName, int credits, int totalStudents, int capacity) {
+public record LectureViewResponse (Long id, String lectureTitle, String professorName, int credits, long totalStudents, int capacity) {
 }
